Revert "feat: add component template and apply to subsystem"
This reverts commit bd9deaeb8dcf3c79a4f8f40fdb32bb961ef7d174.
